Version 0.2; Thesaurus class and Language help
Thesaurus can generate variations on sentences with synonyms configured
as you tell it, or can use placeholder "categories" and fill them with a
word from a list of thematically-related terms. It uses the small
GapShuffler class to try to keep a distance between any repeat
occurences of a word. Thesaurus can also generate procedural nation
names, which should work well as placeholders, but simultaneously
generate a language for further usage that should be similar to any
non-English parts of the nation's name.

You can now generate languages by giving a seed as a long to Language
instead of specifically giving an RNG, which should help compatibility
with tools that use their own random number generators.