Major improvements from SquidLib updates, +
Plus a fix to a bug that seems to have been found first in Gabber;
certain sanity checks could cause non-infinite but lengthy loops to
occur before finally finding a valid word and moving on. This was a flaw
in the default sanity checks, and has been fixed. Other than that,
Nahuatl has been added, there are some improvements to Norse, and a few
other possible bugs are fixed.