Big chunk of updates from SquidLib, fix freeze bug
This adds many new languages from SquidLib's FakeLanguageGen class, with
the code altered somewhat to fit the reduced codebase of Gabber. The
registry code has changed to be a bit more useful, tracking languages
and their names in arrays in Language as well as using the existing
registry (though as a LinkedHashMap of String name to Language instead
of an ArrayList of Language). Some GWT issues have been fixed by using
slower but more accurate and safe long values instead of taking a gamble
by using int on GWT (where it is effectively a Java double feigning some
of int behavior). A bug where a suffix might be used in Translator that
would never be able to pass the additional checks for vulgarity that
Translator performs, which resulted in an infinite loop of attempts to
regenerate a word, has been identified and fixed.